Untangle linking from installing
Installing now means only that a formula runs its install routines and files
should now exist in the cellar for that version. To make the formula available
for use it should then be linked.

This was done to allow installing of formula inside of other formula without
creating link steps if they are not desired. One example is the git formula
which wants to install man pages and html docs using the formula framework but
there is not an actual formula linked from the cellar.